move Bomb class to separate file
