+++ models.py
@@ -1,35 +1,6 @@
-from typing import List
-
 import pygame as pg
 
 from settings import Settings
-
-
-class Explosion(pg.sprite.Sprite):
-    """An explosion. Hopefully the Alien and not the player!"""
-
-    defaultlife = 12
-    animcycle = 3
-    images: List[pg.Surface] = []
-
-    def __init__(self, actor, *groups):
-        pg.sprite.Sprite.__init__(self, *groups)
-        self.image = self.images[0]
-        self.rect = self.image.get_rect(center=actor.rect.center)
-        self.life = self.defaultlife
-
-    def update(self):
-        """called every time around the game loop.
-
-        Show the explosion surface for 'defaultlife'.
-        Every game tick(update), we decrease the 'life'.
-
-        Also we animate the explosion.
-        """
-        self.life = self.life - 1
-        self.image = self.images[self.life // self.animcycle % 2]
-        if self.life <= 0:
-            self.kill()
 
 
 class Score(pg.sprite.Sprite):
